Deleting IzPack BI installer spike (MIFOSBI-56).
IzPack looks like it makes a fine installer, but Sumit and I decided to
look at something much simpler (ie: shell scripts/etc) to install
reports and ETL jobs until we know we need/want the features of
something like IzPack.

Update on that: we may actually use a deb package for Pentaho.

http://mifosforge.jira.com/browse/MIFOSBI-56