Removed WriteToLog step "Got an error on Insert/Update".
It wasn't working as expected. Instead of logging error'd rows from
earlier steps, it was swallowing the error. If logging is turned up to
"detailed", the results are slightly better, but the job should crash if
ppi scores are null, for instance.

I also added a step to remove the scoring_sql field after the Dynamic
SQL step, so row data is cleaner after that point. This is helpful
during debugging.